Added Communication Between services
When a new workspace gets created in the workspace service a new db schema gets applied to the authentication service. Also a password client gets created in the schema of the workspace. This gets communicated back using jobs to the other services. Implemented service classes for database interactions in the authentication repo